+++ packages/frameworks/vue/src/tags-input/tags-input-context.ts
@@ -1,7 +1,7 @@
 import { createContext } from '../context'
 import type { UseTagsInputReturn } from './use-tags-input'
 
-export type TagsInputContext = UseTagsInputReturn
+export interface TagsInputContext extends UseTagsInputReturn {}
 
 export const [TagsInputProvider, useTagsInputContext] =
   createContext<TagsInputContext>('TagsInputContext')