+++ packages/frameworks/vue/src/tags-input/tags-input-input.tsx
@@ -1,11 +1,10 @@
 import { computed, defineComponent } from 'vue'
 import { ark, type HTMLArkProps } from '../factory'
-import type { ComponentWithProps } from '../utils'
 import { useTagsInputContext } from './tags-input-context'
 
-export type TagsInputInputProps = HTMLArkProps<'input'>
+export interface TagsInputInputProps extends HTMLArkProps<'input'> {}
 
-export const TagsInputInput: ComponentWithProps<TagsInputInputProps> = defineComponent({
+export const TagsInputInput = defineComponent({
   name: 'TagsInputInput',
   setup(_, { attrs }) {
     const api = useTagsInputContext()